코테/카카오연습용/requirements.txt
asgiref==3.2.10
boto3==1.14.7
botocore==1.17.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
